requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.32
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0


